# Improve performance of no ops