This PR fixes #94 and contains the following:

- only import ``instances.py`` and ``requests`` when needed (speedup of ``~70ms`` on my machine, run time from ``~220ms`` to ``~150ms``)
- caching of validation results in ``validation.cache`` to avoid importing``jsonschema``:
    - if the cache entry matches the modification time stamp of the file, the file will not be validated (thus ``jsonschema`` will not be imported, additional speedup of ``~80ms``)
    - otherwise, the file will be validated (up to ``~5ms`` slowdown compared to validation without caching, caused by the additional operations involved in the caching process)
        - if the validation is succesful, the cache entry will be updated
        - otherwise, the cache entry will not be updated (thus, the file will be validated the next time ``simex`` is called)