@duylethanh 2021-01-23
Sorry, for all the force pushing. This should be the last for now. I forgot to mention that I refactored ``base.config_for_dir()``, which lead to the removal of ``base.read_and_validate_setup()``. As I'd like to acces the ``basedir`` in ``util.validate_setup_file()`` I changed the parameters it receives, explicitly stating ``basedir``. This in turn, made ``base.read_and_validate_setup()`` unnecessary, as it only joined the ``basedir`` with the ``setup_file``.

PR ready to be reviewed.