@duylethanh 2021-01-31
I made the requested changes. I also had to make a few adjustments to the validation unit test:

- remove the assertion for ``instdir`` in the returned ``dict``, as this was moved to the ``Config`` constructor
- delete ``validation.cache`` files before running the validation, as there could already be a cache file from previous (other) tests

PR ready to be reviewed.