@avdgrinten 2021-02-02
Can you clarify why we remove the `validation.cache` file in tests? To ensure that the test re-validate even if the file was validated correctly? Or is there any correctness issue without removing the file?