@duylethanh 2021-02-03
I fixed the merge conflicts. I also added the following commit 4d113c5:

- added ``util.read_json_file()``, which returns a json-dict and the last modification timestamp of the file
- ``_validate_dict()`` in ``util.validate_setup_file()`` now takes a json-dict as parameter (instead of a ``jsonschema`` schema path)
- setup files will also be validated if the corresponding schema was modified

Also:

> 
> 
> Can you clarify why we remove the `validation.cache` file in tests? To ensure that the test re-validate even if the file was validated correctly? Or is there any correctness issue without removing the file?

After thinking about it, the removal of the cache should not be necessary.  If an ``experiments.yml`` was already validated successfully (because a previous test used it) it shouldnt be re-validated again.

If everything is fine with commit 4d113c5, this PR should be ready to be merged.